sync with main repository

+++ src/service.h
@@ -292,7 +292,7 @@ typedef struct service {
   /**
    * Channel info
    */
-  int         (*s_channel_number) (struct service *);
+  int64_t     (*s_channel_number) (struct service *);
   const char *(*s_channel_name)   (struct service *);
   const char *(*s_provider_name)  (struct service *);
 
@@ -554,6 +554,6 @@ void sort_elementary_streams(service_t *t);
 
 const char *service_get_channel_name (service_t *s);
 const char *service_get_full_channel_name (service_t *s);
-int         service_get_channel_number (service_t *s);
+int64_t     service_get_channel_number (service_t *s);
 
 #endif // SERVICE_H__
